requirements.txt
pywebview==4.4.1
# Add other dependencies as needed
